Added the Midi Instrument example
This allows you to pick a MIDI instrument.  This differs from Piano in that soundfonts are not loaded from the local server but remotely from Benjamin Gleitz's github project where they originated.

Also, amended the soundfont-ports package to enable the remote loading.  This will eventually be backported to the elm-soundfont-ports project.